Use both easymotion & labelless vim-sneak.
They both have their uses, I suppose. vim-sneak is nice and minimal, but
I notice that I appreciate the visual feedback from easymotion, and
vim-sneak's label mode is not quite the same since there is a moment of
uncertainty when I am uncertaintain that I might have to press a third
key.