Swap symbols around.
Having `*` as a standard key is more helpful than `~`, because it's used
in normal typing for emphasis. Then `@` needs to move over so that we
can still use it in number mode. `$` is often used, so should go to the
left side. I think I can sacrifice the `.`. `+` above `-` makes sense
and mirrors `pgup`/`pgdown` in the other layer.